# Compute real matrix logarithm and matrix square root using real arithmetic