--- a/stdlib/LinearAlgebra/src/dense.jl
+++ b/stdlib/LinearAlgebra/src/dense.jl
@@ -803,6 +803,21 @@ function sqrt(A::StridedMatrix{<:Complex})
     end
 end
 
+Base.@propagate_inbounds function _sqrt_real_2x2!(R, A)
+    a11, a21, a12, a22 = A[1, 1], A[2, 1], A[1, 2], A[2, 2]
+    θ = (a11 + a22) / 2
+    μ = sqrt(-(a11 - a22)^2 - 4 * a21 * a12) / 2
+    t = sqrt((abs(θ) + sqrt(θ^2 + μ^2)) / 2)
+    α = θ > 0 ? t : μ / 2t
+    c = 2α
+    d = α - θ / c
+    R[1, 1] = α + (a11 - θ) / c
+    R[2, 1] = a21 / c
+    R[1, 2] = a12 / c
+    R[2, 2] = α + (a22 - θ) / c
+    return R
+end
+
 function inv(A::StridedMatrix{T}) where T
     checksquare(A)
     S = typeof((one(T)*zero(T) + one(T)*zero(T))/one(T))