… stdlib/LinearAlgebra/test/dense.jl
… stdlib/LinearAlgebra/test/triangular.jl