As discussed on Slack (cc @andreasnoack and @baggepinnen), this PR implements the real matrix square root and real matrix logarithm in the following papers, respectively:
- Higham, Nicholas J. (1987) Computing real square roots of a real matrix. Linear Algebra and its Applications, 88-89, 405-430. [eprint](https://www.maths.manchester.ac.uk/~higham/narep/narep89.pdf)
- Al-Mohy, Awad H. and Higham, Nicholas J. and Relton, Samuel D. (2013) Computing the Frechet Derivative of the Matrix Logarithm and Estimating the Condition Number. SIAM J. Sci. Comput., 35 (4). C394 -C410. ISSN 1095-7197. [eprint](http://eprints.maths.manchester.ac.uk/2015/)

These algorithms are both recommended in:
- Higham, Nicholas J. and Hopkins, Edvin (2020) A Catalogue of Software for Matrix Functions. Version 3.0. [eprint](http://eprints.maths.manchester.ac.uk/2754/)

This was proposed in https://github.com/JuliaLang/julia/issues/5840 and discussed for `sqrt` in https://github.com/JuliaLang/julia/issues/4006.

The result is that whenever a real logarithm or square root exists, it is returned. This both speeds up the algorithms (see benchmarks below) and prevents error from accumulating in the imaginary parts of the outputs. To avoid heavy code duplication, the PR refactors the `log` and `sqrt` of upper triangular matrices to more generally compute `log` and `sqrt` of upper quasi-triangular matrices.

The changes:
- Previously a complex result was always returned for `sqrt(A::Matrix{<:Real})` and `log(A::Matrix{<:Real})` unless `A` was symmetric with positive eigenvalues, in which case a real result was returned. Now a real result is returned whenever the eigenvalues are positive.
- For a real matrix `A` (with `Real` or `Complex` eltype), if a real matrix square root or real matrix logarithm exists, then it is computed using entirely real arithmetic. If the eltype of `A` is `Complex`, then a complex result is returned, where the imaginary part is exactly zero. Previously the function would be computed using complex arithmetic, resulting in imaginary parts that were not exactly zero.
- `log(::UnitUpperTriangular)` and `log(::UnitLowerTriangular)` are now defined.

Issues fixed along the way:
- For `Float32` and `ComplexF32` inputs, `log` always promoted to the `Float64` analogs. It no longer does.
- The inferred return type of `log(::Matrix{ComplexF64})` was `Any`. It is now `Matrix{ComplexF64}`.
- For `A::UpperTriangular{<:Real}` with negative eigenvalues, `sqrt(A)` returned a complex result, while `log(A)` would error. `log(A)` now returns a complex result.

## Example

Before:
```julia
julia> log(exp([1 -2; 1 1]))
2×2 Matrix{ComplexF64}:
 1.0+3.33067e-16im  -2.0+1.66533e-16im
 1.0-3.33067e-16im   1.0+1.66533e-16im
```

This PR:
```julia
julia> log(exp([1 -2; 1 1]))
2×2 Matrix{Float64}:
 1.0  -2.0
 1.0   1.0
```

## Basic Benchmarks

Here are some quick and dirty benchmarks on my (very old) machine.

### Matrix square root

```julia
using LinearAlgebra, BenchmarkTools

n = 20
Artor = randn(n, n)^2;
Actor = complex(Artor);
Artoc = randn(n, n);
Actoc = randn(ComplexF64, n, n);

println("Dense matrices:")
for (name, mat) in (("ℝ → ℝ", Artor), ("ℂ → ℝ", Actor), ("ℝ → ℂ", Artoc), ("ℂ → ℂ", Actoc))
    @assert sqrt(mat)^2 ≈ mat
    println(name)
    @btime sqrt($mat);
end

Urtor = UpperTriangular(randn(n, n))^2;
Uctor = complex(Urtor);
Urtoc = UpperTriangular(randn(n, n));
Uctoc = UpperTriangular(randn(ComplexF64, n, n));

println("\nUpper triangular matrices:")
for (name, mat) in (("ℝ → ℝ", Urtor), ("ℂ → ℝ", Uctor), ("ℝ → ℂ", Urtoc), ("ℂ → ℂ", Uctoc))
    @assert sqrt(mat)^2 ≈ mat
    println(name)
    @btime sqrt($mat);
end
```

Before:
```julia
Dense matrices:
ℝ → ℝ
  266.140 μs (10 allocations: 105.59 KiB)
ℂ → ℝ
  265.059 μs (9 allocations: 99.22 KiB)
ℝ → ℂ
  310.661 μs (10 allocations: 105.59 KiB)
ℂ → ℂ
  320.669 μs (9 allocations: 99.22 KiB)

Upper triangular matrices:
ℝ → ℝ
  3.123 μs (2 allocations: 3.27 KiB)
ℂ → ℝ
  7.821 μs (1 allocation: 6.38 KiB)
ℝ → ℂ
  6.819 μs (2 allocations: 6.39 KiB)
ℂ → ℂ
  6.669 μs (1 allocation: 6.38 KiB)
```

This PR:
```julia
Dense matrices:
ℝ → ℝ
  110.562 μs (120 allocations: 58.70 KiB)
ℂ → ℝ
  111.417 μs (122 allocations: 68.33 KiB)
ℝ → ℂ
  154.114 μs (22 allocations: 159.44 KiB)
ℂ → ℂ
  321.852 μs (10 allocations: 99.23 KiB)

Upper triangular matrices:
ℝ → ℝ
  3.072 μs (2 allocations: 3.27 KiB)
ℂ → ℝ
  5.062 μs (4 allocations: 12.89 KiB)
ℝ → ℂ
  5.771 μs (2 allocations: 6.39 KiB)
ℂ → ℂ
  5.580 μs (2 allocations: 6.39 KiB)
```

### Matrix logarithm

```julia
using LinearAlgebra, BenchmarkTools
n = 20
Artor = exp(randn(n, n));
Actor = complex(Artor);
Artoc = randn(n, n);
Actoc = randn(ComplexF64, n, n);

Urtor = UpperTriangular(exp(triu(randn(n, n))));
Uctor = complex(Urtor);
Urtoc = UpperTriangular(randn(n, n));
Uctoc = UpperTriangular(randn(ComplexF64, n, n));

println("Dense matrices:")
for (name, mat) in (("ℝ → ℝ", Artor), ("ℂ → ℝ", Actor), ("ℝ → ℂ", Artoc), ("ℂ → ℂ", Actoc))
    @assert exp(log(mat)) ≈ mat
    println(name)
    @btime log($mat);
end

println("\nUpper triangular matrices:")
for (name, mat) in (("ℝ → ℝ", Urtor), ("ℂ → ℝ", Uctor), ("ℝ → ℂ", Urtoc), ("ℂ → ℂ", Uctoc))
    println(name)
    try
        @assert exp(Matrix(log(mat))) ≈ mat
        @btime log($mat);
    catch
        println("  Error")
    end
end
```

Before:
```julia
Dense matrices:
ℝ → ℝ
  335.860 μs (92 allocations: 490.84 KiB)
ℂ → ℝ
  342.864 μs (102 allocations: 494.31 KiB)
ℝ → ℂ
  417.632 μs (97 allocations: 495.61 KiB)
ℂ → ℂ
  560.571 μs (85 allocations: 423.34 KiB)

Upper triangular matrices:
ℝ → ℝ
  103.731 μs (123 allocations: 169.84 KiB)
ℂ → ℝ
  190.273 μs (70 allocations: 325.05 KiB)
ℝ → ℂ
  Error
ℂ → ℂ
  352.777 μs (101 allocations: 464.41 KiB)
```

This PR:
```julia
Dense matrices:
ℝ → ℝ
  254.740 μs (506 allocations: 273.17 KiB)
ℂ → ℝ
  254.052 μs (508 allocations: 282.80 KiB)
ℝ → ℂ
  402.231 μs (103 allocations: 489.34 KiB)
ℂ → ℂ
  553.679 μs (87 allocations: 423.36 KiB)

Upper triangular matrices:
ℝ → ℝ
  99.269 μs (124 allocations: 170.08 KiB)
ℂ → ℝ
  101.995 μs (124 allocations: 176.59 KiB)
ℝ → ℂ
  407.174 μs (231 allocations: 556.72 KiB)
ℂ → ℂ
  346.416 μs (192 allocations: 459.47 KiB)
```