@sethaxen 2021-03-10
> Do you have an idea if you exercise all the branches and functions with the current tests?

I'm pretty sure I cover all of the new branches, but there may be old branches that were not covered. I'll double check.

> I'm also a bit surprised about the number of allocations. I noticed one place (the rdiv) that could trigger some allocations but I'm also wondering if any of the helper functions could have some type stability issues. Do you know why the allocations happen?

I can explain a few of the allocations. Real `sqrt` will call `LAPACK.trsyl!` for all off-diagonal 2x2 blocks, and this function I believe allocates a single-element array. In other cases, copies are made of the inputs or outputs to constrain the types so that the helpers will be type-stable, but I don't think these explain the high increase in allocations. Whenever I made a fix to decrease the allocations, I saw runtime increase. I'm still trying to work this out. I'm only aware of allocations as creation of new arrays, which should not happen due to the views. How would type-instability increase allocations?