@andreasnoack 2021-03-19
The current version is definitely already an improvement so if you want to stop here then I think it's okay. However, I think it's a shame that there are so many allocations. Ideally, the LAPACK wrappers would allow reuse of memory as proposed in https://github.com/JuliaLang/julia/issues/16263 and it should be fairly straightforward to modify the `trsyl!` method to take https://github.com/JuliaLang/julia/blob/efad4e3134540c7a45d385b047f145cb0da6d385/stdlib/LinearAlgebra/src/lapack.jl#L6460-L6461 as arguments while giving them default values similar to their current values.