@sethaxen 2021-03-19
> It should probably be
> ```julia
> scale = Ref{$relty}()
> ```
> 
> though

It looks like this change alone eliminated ~half of the allocations. The remaining ones come from `checksquare(A, B)`, which oddly allocates a vector with the sizes of `A` and `B` instead of returning a 2-tuple (this is the documented behavior). Since it does nothing fancier than essentially `checksquare` on each argument, I've changed `trsyl!` to just do that. I did not need to modify the signature of `trsyl!`. With these changes, `_sqrt_quasitriu!` is completely non-allocating.

I also eliminated excess allocations when computing the degree of the Padé approximant. Now the same 3 allocations are reused for all steps. This also cut ~1/3 of `_log_quasitriu!`s allocations.

## Updated benchmarks

### Matrix square root
All remaining allocations come from copies or the Schur decomposition.
```julia
Dense matrices:
ℝ → ℝ
  115.852 μs (12 allocations: 50.83 KiB)
ℂ → ℝ
  117.443 μs (14 allocations: 60.45 KiB)
ℝ → ℂ
  145.638 μs (21 allocations: 153.06 KiB)
ℂ → ℂ
  316.831 μs (10 allocations: 99.23 KiB)

Upper triangular matrices:
ℝ → ℝ
  3.050 μs (2 allocations: 3.27 KiB)
ℂ → ℝ
  5.179 μs (4 allocations: 12.89 KiB)
ℝ → ℂ
  5.641 μs (2 allocations: 6.39 KiB)
ℂ → ℂ
  5.529 μs (2 allocations: 6.39 KiB)
```

### Matrix logarithm
```julia
Dense matrices:
ℝ → ℝ
  220.330 μs (42 allocations: 96.14 KiB)
ℂ → ℝ
  221.420 μs (44 allocations: 105.77 KiB)
ℝ → ℂ
  330.091 μs (48 allocations: 228.39 KiB)
ℂ → ℂ
  501.972 μs (37 allocations: 174.56 KiB)

Upper triangular matrices:
ℝ → ℝ
  76.920 μs (32 allocations: 48.75 KiB)
ℂ → ℝ
  78.498 μs (32 allocations: 55.27 KiB)
ℝ → ℂ
  249.985 μs (31 allocations: 81.97 KiB)
ℂ → ℂ
  295.925 μs (29 allocations: 81.72 KiB)
```