@andreasnoack 2021-03-20
> It looks like this change alone eliminated ~half of the allocations.

Great. This is of course a much cleaner solution. I need to update my mental model of the compiler to something post Julia 0.5.

> The remaining ones come from `checksquare(A, B)`, which oddly allocates a vector

Yes, that's unfortunate. See https://github.com/JuliaLang/julia/issues/28897